Cleaning up compatiblity issues, reimplementing fuzzy name search using a library
