+++ player_properties.txt
