+++ src/main/java/Midfielder.java
@@ -1,6 +1,14 @@
+import java.util.HashMap;
+
 public class Midfielder extends Player{
-    public Midfielder(){
-        setName("Midfielder");
-        setAge(32);
+    public Midfielder(String name, int age, double height, double weight, String team, int rating,
+                      double value, String position, HashMap<String, Integer> skills)
+    {
+        super(name, age, height, weight, team, rating, value, position, skills);
+
     }
+
+    @Override
+    public void displayPos() {
+        System.out.println("This player's pos is " + super.getPosition());}
 }
